requirements.txt
telethon
python-dotenv
pandas
psycopg2-binary
sqlalchemy
dbt-postgres
ultralytics
fastapi
uvicorn
dagster
dagster-webserver